requirements.txt
Django==5.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.7
python-dotenv==1.0.0
colorama==0.4.6
loguru==0.7.2
Pillow==10.0.0
gunicorn==21.2.0
